Requirement: #966 Requirement: #965
Merging changes for Hibernate 1.2 and Spring.NET 1.1 upgrades into the Trunk.

+++ Enterprise/Authentication/AuthenticationService.cs
@@ -54,7 +54,7 @@ namespace ClearCanvas.Enterprise.Authentication
             UserSearchCriteria where = new UserSearchCriteria();
             where.UserName.EqualTo(userName);
 
-            int count = PersistenceContext.GetBroker<IUserBroker>().Count(where);
+            long count = PersistenceContext.GetBroker<IUserBroker>().Count(where);
             return count == 1;
         }
