+++ Enterprise/Core/IEntityBroker.cs
@@ -152,14 +152,14 @@ namespace ClearCanvas.Enterprise.Core
         /// </summary>
         /// <param name="criteria"></param>
         /// <returns></returns>
-        int Count(TSearchCriteria criteria);
+        long Count(TSearchCriteria criteria);
 
         /// <summary>
         /// Returns the number of entities matching any of the specified criteria (the criteria are combined using OR).
         /// </summary>
         /// <param name="criteria"></param>
         /// <returns></returns>
-        int Count(TSearchCriteria[] criteria);
+        long Count(TSearchCriteria[] criteria);
 
 
         /// <summary>
