+++ Healthcare/Hibernate/Brokers/RegistrationWorklistBroker.cs
@@ -116,7 +116,7 @@ namespace ClearCanvas.Healthcare.Hibernate.Brokers
         private int GetWorklistCount(string hqlQuery, List<QueryParameter> parameters)
         {
             IList list = DoQuery(hqlQuery, parameters);
-            return (int)list[0];
+            return (int)(long)list[0];
         }
 
         private IList DoQuery(string hqlQuery, List<QueryParameter> parameters)
@@ -369,7 +369,7 @@ namespace ClearCanvas.Healthcare.Hibernate.Brokers
             AddMainQueryParameters(parameters);
 
             IList list = DoQuery(hqlQuery, parameters);
-            return (int)list[0];
+            return (int)(long)list[0];
         }
 
         public IList<Order> GetOrdersForCancel(Patient patient)
@@ -420,7 +420,7 @@ namespace ClearCanvas.Healthcare.Hibernate.Brokers
             AddSubQueryParameters(parameters);
 
             IList list = DoQuery(hqlQuery, parameters);
-            return (int)list[0];
+            return (int)(long)list[0];
         }
 
         #endregion
