+++ Healthcare/Hibernate/Brokers/ReportingWorklistBroker.cs
@@ -127,7 +127,7 @@ namespace ClearCanvas.Healthcare.Hibernate.Brokers
         private int GetWorklistCount(string hqlQuery, IEnumerable<QueryParameter> parameters)
         {
             IList list = DoQuery(hqlQuery, parameters);
-            return (int)list[0];
+            return (int)(long)list[0];
         }
 
         private IList DoQuery(string hqlQuery, IEnumerable<QueryParameter> parameters)
